pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "photo-dedup"
version = "1.4.0"
description = "Find and remove duplicate photos in your Google Photos backup (or any photo directory)."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
heic = ["pillow-heif>=0.14.0"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.6.0",
]

[project.scripts]
photo-dedup-scan = "scan:main"
photo-dedup-clean = "clean:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
